refactor(lore): reorganize domain types into lore/domain/
This commit moves MailingList and Patch-related types
from their flat lore/mailing_list/ and lore/patch/
subdirectories into a dedicated lore/domain/ layer.
MessageID is extracted into its own lore/domain/ids.rs
file and re-exported from patch.rs for backward-compatible access.

It also creates empty lore/application/ and
lore/infrastructure/ module stubs to scaffold the
full bounded context structure for upcoming commits.

This commit is part of the architecture refactoring's phase 2.
