+++ src/lore/application/mod.rs
@@ -1,0 +1,1 @@
+// Lore application layer — populated in subsequent commits.
