+++ src/lore/domain/mod.rs
@@ -1,0 +1,3 @@
+pub mod ids;
+pub mod mailing_list;
+pub mod patch;
