+++ src/lore/infrastructure/mod.rs
@@ -1,0 +1,1 @@
+// Lore infrastructure adapters — populated in subsequent commits.
