+++ src/lore/mod.rs
@@ -1,4 +1,5 @@
+pub mod application;
+pub mod domain;
+pub mod infrastructure;
 pub mod lore_api_client;
 pub mod lore_session;
-pub mod mailing_list;
-pub mod patch;
